# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loup — household services marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List household bookings
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [active, upcoming, past, all]
          description: Filter bookings by lifecycle scope (default all)
      responses:
        "200":
          description: Bookings, soonest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Booking detail with status timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Reschedule or cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /bookings/{id}/advance:
    post:
      operationId: advanceBooking
      tags: [bookings]
      summary: Advance a live booking to its next status (demo control)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking after advancing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
        "400":
          description: Booking cannot advance further
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /bookings/{id}/messages:
    get:
      operationId: listBookingMessages
      tags: [bookings]
      summary: Booking-scoped chat messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendBookingMessage
      tags: [bookings]
      summary: Send a chat message on a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /billing/pay:
    post:
      operationId: payBillingStatement
      tags: [billing]
      summary: Settle the current consolidated bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Statement after payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillStatement"
        "400":
          description: Nothing to pay or invalid method
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: integer
        providerName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        categoryName:
          type: string
        categoryIcon:
          type: string
        memberId:
          type: integer
        memberName:
          type: string
        addressId:
          type: integer
        addressLabel:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/BookingStatus"
        priceEstimate:
          type: number
        instructions:
          type: ["string", "null"]
        etaMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - providerId
        - providerName
        - serviceId
        - serviceName
        - categoryName
        - categoryIcon
        - memberId
        - memberName
        - addressId
        - addressLabel
        - scheduledAt
        - status
        - priceEstimate
        - instructions
        - etaMinutes
        - createdAt
    BookingStatus:
      type: string
      enum:
        - pending
        - accepted
        - confirmed
        - en_route
        - arrived
        - in_progress
        - completed
        - cancelled
        - rejected
        - disputed
    BookingEvent:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        status:
          $ref: "#/components/schemas/BookingStatus"
        note:
          type: string
        occurredAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - status
        - note
        - occurredAt
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/BookingEvent"
            hasReview:
              type: boolean
          required:
            - events
            - hasReview
    BookingInput:
      type: object
      properties:
        serviceId:
          type: integer
        addressId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        instructions:
          type: string
        allowanceContribution:
          type: number
          description: How much of the employee's benefit allowance to apply (0 = pay full price, max = available balance up to service price)
        memberId:
          type: integer
          description: Which household member this booking is for (defaults to current user)
      required:
        - serviceId
        - addressId
        - scheduledAt
...
    BookingUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [cancelled]
    Message:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        sender:
          type: string
          enum: [member, provider]
        senderName:
          type: string
        body:
          type: string
        sentAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - sender
        - senderName
        - body
        - sentAt
    MessageInput:
      type: object
      properties:
        body:
          type: string
          minLength: 1
      required:
        - body
...
    BillItem:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        serviceName:
          type: string
        providerName:
          type: string
        memberName:
          type: string
        categoryName:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - serviceName
        - providerName
        - memberName
        - categoryName
        - amount
        - date
    MemberSpend:
      type: object
      properties:
        memberId:
          type: integer
        memberName:
          type: string
        initials:
          type: string
        amount:
          type: number
      required:
        - memberId
        - memberName
        - initials
        - amount
    CategorySpend:
      type: object
      properties:
        categoryName:
          type: string
        amount:
          type: number
      required:
        - categoryName
        - amount
    BillStatement:
      type: object
      properties:
        id:
          type: integer
        month:
          type: string
          description: Human label, e.g. "August 2026"
        status:
          type: string
          enum: [open, paid]
        total:
          type: number
        paidAt:
          type: ["string", "null"]
          format: date-time
        paidWith:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillItem"
        byMember:
          type: array
          items:
            $ref: "#/components/schemas/MemberSpend"
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySpend"
      required:
        - id
        - month
        - status
        - total
        - paidAt
        - paidWith
        - items
        - byMember
        - byCategory
...
    PaymentInput:
      type: object
      properties:
        paymentMethodId:
          type: integer
      required:
        - paymentMethodId
...